requirements.txt
fastapi==0.95.0
pydantic>=1.8.0,<2.0.0
uvicorn==0.21.1
sqlalchemy==2.0.8
pymysql==1.0.3
python-dotenv==1.0.0
torch==2.3.0
boto3==1.26.160
numpy==1.21.1
pandas==1.3.5
scipy==1.5.2
tqdm==4.62.3
scikit-learn==1.0.1